+++ src/View/templates/error-page.php
@@ -37,13 +37,10 @@
     </nav>
 
     <?php
-        $class_name = $this->info_error_exception['namespace_exception'];
-        $exception = new \ReflectionClass($class_name);
+        $this->solution = new ModernPHPException\Solution();
     ?>
 
-    <?php if (method_exists($exception, "getSolution")) : ?>
-        <?php $exception->getMethod('getSolution')->invoke($class_name); ?>
-
+    <?php if ($this->solution->getTitle() != '') : ?>
         <nav class="navbar navbar-expand-lg navbar-light bg-custom-2 p-3">
             <div class="container-fluid">
                 <div>